pyproject.toml
[project]
name = "posttrend"
version = "0.1.0"
description = "posttrend using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "langchain-community>=0.3.29",
    "langchain-groq",
    "streamlit>=1.49.1",
    "tavily-python>=0.7.12",
]

[project.scripts]
posttrend = "posttrend.main:run"
run_crew = "posttrend.main:run"
train = "posttrend.main:train"
replay = "posttrend.main:replay"
test = "posttrend.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
